Simplify, standardize and add splitting and merging
This is a pretty hefty set of changes.  The big thing was to
remove many of the parameters from the input and effectively hardcode
them in the pipeline.  This is because the pipeline will be run with
the appropriate docker image so all the paths and parameters should
be fixed.  Also, we want to always run the same set of pieces so
some all the flags were hardcoded.

We've also removed runtime parameters that were considered non-standard.
The Cromwell configuration should determine those based on the other parameters
or some other site logic.

This add tasks to do the splitting and merging.  The exact list of files to
merge may be incomplete.